Changed Group.member functions to BaseGroup.member
